Car 37

See merge request xhorsky/pv-168-project-share-car-rider!27

+++ src/main/java/cz/muni/fi/pv168/project/business/model/CarRide.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 public class CarRide extends Template {
     private LocalDateTime date;
 
-    public CarRide(String guid, String Title, String Description, Double Distance, double FuelConsumption, int CostOfFuelPerLitre, int NumberOfPassengers, double commission, LocalDateTime Date, Category Category) {
+    public CarRide(String guid, String Title, String Description, Double Distance, double FuelConsumption, double CostOfFuelPerLitre, int NumberOfPassengers, double commission, LocalDateTime Date, Category Category) {
         super(guid, Title, Description, Distance, FuelConsumption, CostOfFuelPerLitre, NumberOfPassengers, commission, Category);
         setDate(Date);
     }
